Update following review:
- disable Rest endpoint
- use a `BaseContentItem` model as parent class for `MetadataResource`
  and `ContentAPIItem`
- use `fields.Keyword` when `not_analyzed` is used in legacy resource.
- added default values for many fields
- removed some custom models in favor of existing ones from `content_api.items.model`
