+++ content_api/items/model.py
@@ -1,95 +1,12 @@
-from typing import Annotated, Any
-from datetime import datetime
+from typing import Annotated
 from enum import Enum, unique
 
 from pydantic import Field, field_validator
 
-from superdesk.core.resources import ResourceModel, dataclass, Dataclass, fields, validators, ModelWithVersions
-from superdesk.utc import utcnow
+from superdesk.core.resources import ResourceModel, fields, ModelWithVersions
+from superdesk.types.base import CVItemWithCode, ContentType, Place
 
 
-ContentAssociation = Annotated[
-    dict[str, Any],
-    fields.elastic_mapping(
-        {
-            "dynamic": False,
-            "properties": {
-                "featuremedia": {
-                    "dynamic": False,
-                    "properties": {
-                        "_id": {"type": "keyword"},
-                        "guid": {"type": "keyword"},
-                        "unique_id": {"type": "integer"},
-                    },
-                },
-            },
-        }
-    ),
-]
-
-
-@dataclass
-class CVItem(Dataclass):
-    qcode: fields.Keyword
-    name: fields.Keyword
-    scheme: fields.Keyword | None = None
-    schema: fields.Keyword | None = None
-
-
-@dataclass
-class CVItemWithCode(Dataclass):
-    code: fields.Keyword
-    name: fields.Keyword
-    schema: fields.Keyword | None = None
-    scheme: fields.Keyword | None = None
-
-
-@dataclass
-class Place(Dataclass):
-    scheme: fields.Keyword | None = None
-    qcode: fields.Keyword | None = None
-    code: fields.Keyword | None = None
-    name: fields.Keyword | None = None
-    locality: fields.Keyword | None = None
-    state: fields.Keyword | None = None
-    country: fields.Keyword | None = None
-    world_region: fields.Keyword | None = None
-    locality_code: fields.Keyword | None = None
-    state_code: fields.Keyword | None = None
-    country_code: fields.Keyword | None = None
-    world_region_code: fields.Keyword | None = None
-    feature_class: fields.Keyword | None = None
-    location: fields.Geopoint | None = None
-    rel: fields.Keyword | None = None
-
-
-@dataclass
-class Annotation(Dataclass):
-    id: int
-    type: fields.Keyword
-    body: fields.Keyword
-
-
-@dataclass
-class ContentAuthor(Dataclass):
-    uri: fields.Keyword | None = None
-    parent: fields.Keyword | None = None
-    name: fields.TextWithKeyword | None = None
-    role: fields.Keyword | None = None
-    jobtitle: dict | None = None
-    sub_label: fields.TextWithKeyword | None = None
-    biography: str | None = None
-    code: fields.Keyword | None = None
-
-
-@dataclass
-class ContentReference(Dataclass):
-    id: Annotated[fields.Keyword, Field(alias="_id")]
-    key: fields.Keyword | None = None
-    uri: fields.Keyword | None = None
-    guid: fields.Keyword | None = None
-    type: fields.Keyword | None = None
-    source: fields.Keyword | None = None
 
 
 @unique
@@ -99,57 +16,24 @@ class PubStatusType(str, Enum):
     CANCELLED = "canceled"
 
 
-@unique
-class ContentType(str, Enum):
-    TEXT = "text"
-    PREFORMATTED = "preformatted"
-    AUDIO = "audio"
-    VIDEO = "video"
-    PICTURE = "picture"
-    GRAPHIC = "graphic"
-    COMPOSITE = "composite"
-
 
 class ContentAPIItem(ResourceModel, ModelWithVersions):
-    id: Annotated[str, Field(alias="_id")]
-    guid: str | None = None
-    associations: ContentAssociation | None = None
-    anpa_category: list[CVItem] = Field(default_factory=list)
-    body_html: fields.HTML | None = None
-    body_text: str | None = None
-    byline: str | None = None
-    copyrightnotice: Annotated[str | None, fields.not_indexed()] = None
-    copyrightholder: str | None = None
     description_html: str | None = None
-    description_text: str | None = None
-    headline: fields.HTML | None = None
-    language: fields.Keyword | None = None
     located: str | None = None
     mimetype: fields.Keyword | None = None
     organization: list[dict] = Field(default_factory=list)
     person: list[dict] = Field(default_factory=list)
     place: list[Place] = Field(default_factory=list)
     profile: str | None = None
-    pubstatus: Annotated[PubStatusType | None, fields.keyword_mapping()] = None
     renditions: dict | None = None
     service: list[CVItemWithCode] = Field(default_factory=list)
-    slugline: str | None = None
     source: fields.Keyword | None = None
-    subject: list[CVItemWithCode] = Field(default_factory=list)
     keywords: list[fields.HTML] = Field(default_factory=list)
     anpa_take_key: str | None = None
 
     content_type: Annotated[ContentType, fields.keyword_mapping(), Field(alias="type")] = ContentType.TEXT
 
-    urgency: int | None = None
-    priority: int | None = None
-    uri: Annotated[fields.Keyword | None, validators.validate_iunique_value_async("items", "uri")] = None
-    usageterms: str | None = None
     version: str | None = None
-    versioncreated: datetime = Field(default_factory=utcnow)
-    firstcreated: datetime = Field(default_factory=utcnow)
-    firstpublished: datetime = Field(default_factory=utcnow)
-    embargoed: datetime | None = None
     evolvedfrom: fields.Keyword | None = None
     nextversion: fields.Keyword | None = None
     original_id: fields.Keyword | None = None
@@ -159,11 +43,8 @@ class ContentAPIItem(ResourceModel, ModelWithVersions):
     genre: list[CVItemWithCode] = Field(default_factory=list)
     ancestors: Annotated[list[fields.Keyword], fields.keyword_mapping(), Field(default_factory=list)]
     attachments: list[dict] = Field(default_factory=list)
-    annotations: list[Annotation] = Field(default_factory=list)
 
-    extra: dict | None = None
     extra_items: dict | None = None
-    authors: list[ContentAuthor] = Field(default_factory=list)
     wordcount: int | None = None
     charcount: int | None = None
     readtime: int | None = None
@@ -175,9 +56,10 @@ class ContentAPIItem(ResourceModel, ModelWithVersions):
     agenda_id: fields.Keyword | None = None
     agenda_href: fields.Keyword | None = None
 
-    refs: list[ContentReference] = Field(default_factory=list)
-    expiry: datetime | None = None
 
     @field_validator("version", mode="before")
     def parse_version(cls, value: int | str | None) -> str | None:
         return str(value) if value is not None else None
+
+
+
